Move `ParallelConfig` from `config/__init__.py` to `config/parallel.py`
